requirements.txt
# Core ML
pandas==2.2.1
numpy==1.26.4
scikit-learn==1.4.2
joblib==1.3.2

# API
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.3

# Monitoring
prometheus-client==0.20.0

# Data versioning
dvc==3.42.0



# Utilities
pyyaml==6.0.1
requests==2.31.0
